UDP: connected mode re-implemented
UDPSingleton used to work in connected mode, where it first connected
to the specified remote and then it accepted packets only from that
remote. This causes problems however that when two listeners that
synchronize on a Sync cluster connect at different times then one side
will fail because the send() will return with an ECONN error.

We now have two new semantics:

- unconnected mode: bind socket, wait for the first packet, use the
  source IP and source port as remote, connect back to this remote,
  create a stream

- connected mode: bind socket, wait for the first packet whose remote
  address and port match for the specified remote, connect back to
  this remote, create a stream