Implemented the JSONSocket listener
Idea is that the first packet of a stream (on top of any datagram
transport, like WebSocket or UDP) should be a "header" (a'la HTTP) in
JSON format. This header is then used to update the session metadata,
making it possible to create the impression of a stream semantics with
session setup/header phase for transports that otherwise do not
support this (i.e., UDP).
TODO: implement a JSONSocket cluster and implement a simple
"reliability" semantics for the JSON header (resend after timeout)