Fix bug in Routes API test
I have to modify the add-existing-route test case posData
variable because the name was hardcoded and somehow in the big
npm test have a problem with it.

I think the l7mp object are still exists after API tests, so
could be good if we can find a way to close after each test.