addCluster() has become async
So far addCluster() was synchronous, which is OK for all clusters but
ControllerCluster, because the OpenAPI backend init code is
asynchronous. The async init part was separated out into an
async run() method (the constructor cannot be async), but then
all calls to Cluster.create and addCluster (including all tests) had
to be updated as well.