Convert the add* API endpoints to async
Unfortunately, addCluster calls may take a long time, especially when
adding the Controller cluster that has to parse the OpenAPI spec. This
causes all sorts of weird timeouts in the API tests. Now converting
addCluster to async brings the entire add* path of the API with itself,
since addCluster calls may be embedded addRoute calls (with implicit
cluster defs) which make addRoute also async, and so on.