l7mp version string sanitization
We use git-repo-version to get the version string as 'package.jon:version + git-commit-hash +
git-commit-date'.  If there's no git metadata in the root dir then the latter two are missing in
which case we get a 'null' in the version string: we remove it and remain with the package.json
version in such cases.