JSONSocketCluster rewrite: delegate everything to the transport
So far, JSONSocketCluster has maintained its own loadbalancer on top of the
loadbalancer of the transport cluster, which created all sorts of issues when
loadbalancer.update() was updating the top-level cluster instead of the
transport cluster's loadbalancer. This commit should fix this. No changes
should be visible to the users of JSONSocketCluster.