Implement deleteRuleFromRuleList whislist #1
In recursive mode (recursive=true) we not just remove the rule from
the rulelist but we also delete the corresponding rule as well, also
recursively. Note that this may bring with itself each sessions that
was routed through a cluster that was implicitly created in the
rule. Note further that if deleteRuleFromRuleList was called with a
position as argument then in recursive mode the rule corresponding to
the position will be deleted.