UDP offload: destructively set UDP checksum to 0
This patch solves a bug where incremental checksum update does not work when offloading a proxy
link from localhost, as UDP packets traveling on lo do not seem to contain a valid checksum and so
we update the wrong sequence number. The patch will explicitly set the UDP checksum to zero and set
skb->hash, skb->swhash and skb->l4hash to 0. This seems to magically take care of wrong sequence
numbers